Backports for node 14: node, c-ares, icu, libical
The app under electron 13 requires node 14 native to satisfy yarn and
electron-builder (the actual on-device node is installed by yarn).
That's not in dunfell, but it is in later releases, so backport
- Node 14, with its patches
- c-ares, icu, and libical
And select them using PREFERRED_VERSION stanzas in the distro conf to
make sure the right versions are picked up.