ot3: Add modeset in bootargs via tezi
The Toradex Verdin's DSI graphics output driver doesn't have presence
detection, even when there happens to be an HDMI adapter on it. That
means that nothing can really tell the system when there is or isn't a
display, and that means that the normal methods of
modesetting (configuring the kernel to drive a specific resolution,
refresh rate, color depth, etc) which normally are driven by presence
notification and EDID display communication don't work.

Instead, we need to just tell the kernel what we want. This wouldn't
work for a desktop where a user could plug in whatever monitor they
wanted, but that's not what we have. For now, we'll set it to a
reasonable monitor res.

The way the data actually gets written is at the initial flash of the
board through toradex easy installer. The easy installer is a ramdisk
linux userspace that executes helper scripts bundled into the image,
which means we can take advantage of utilities like uboot firmware
interrogation to write the correct bootargs to the uboot configuration
block in the carveout.

correct tezi wrapup override