ot3: add robot app recipe
This adds a recipe to build from git the Opentrons app for presentation
on an OT3, running under wayland.

It takes advantage of electron-builder's capability to build a directory
with some runnable entry points and the js asar packages stuffed in a
content subdirectory. This dir can then be installed, and is installed,
into a place like /opt/opentrons-app. Some libs have to be removed
because they shadow system libs, and since wayland works via rpc that
includes actual addresses, dynamically linking against local versions of
wayland libs that don't match system libs won't work.

Once we have a runnable app, we can drop a systemd script that brings it
up at boot time. This has to do some gymnastics through a shell script
because it's a good practice to set the XDG_RUNTIME_DIR only if it's not
already set, which isn't really something you can do in a systemd unit.