recipes-robot: bundle python app deps from source
The normal openembedded way to handle dependencies is to make recipes
for the dependencies and mark them as dependencies in the endpoint
application recipes.

For systems that have a lot of dependencies, particularly in a modern
language like python or js where those dependencies frequently change,
this is an immense amount of work that is largely unnecessary. Instead,
we can use the dependency resolution and hashchecking mechanic that the
language libraries (in python, pip) already implement to source the
dependencies from the same data that provides the development
environment: the Pipfile.lock. The pipenv-app-bundle class uses a
utility called micropipenv to read packages, versions, and (whenever we
regenerate locks) hashes from Pipfile.lock and install them directly.

That also means we can remove the recipes that we added to supply the
dependencies that we no longer need.