recipes-robot: move hardware into robot server
opentrons-hardware isn't really its own application, and actually will
end up being a dependency of the robot server eventually. until that
point, we can put it in the newly-actually-functioning
PIPENV_APP_BUNDLE_EXTRAS to install into the robot server directory,
remove the old recipe, and remove it from the image.