pipenv_app_bundle: support poetry
micropipenv, despite the name, happily supports doing its thing based on
pyproject.tomls that use poetry in addition to Pipfiles that use pipenv.
To do this, however, it needs toml, so we need to add a native toml dep;
since we're doing that, we might as well just actually properly add
micropipenv as a dep instead of wgeting it.

Once that happens, we need to be able to actually tell the class to use
poetry, and then everything kind of just works.