recipes-robot: add ot3 firmware python dep
This will be used to source canbus message constant data and structure
for the hardware package, which means it needs to be installed into the
robot server virtualenv environment. We can use the same installation
method as the other python packages (now that we support poetry) as long
as we after the fact override FILES to make sure we're not
doublecounting anything.