recipes-devtools: add python3-setuptools 46.4.0
This overrides the version in openembedded-core to work around
https://github.com/pypa/setuptools/issues/3118 , a mysterious problem
that meant if you have a version in a specific range installed, you
can't install versions outside that range. This version is the
lowest-versioned one outside that range, which we do instead of the
highest-version one because that one can't be installed with the built
in pypi module.