classes/pipenv_app_bundle: set sysconfigdata name
Some python or distutils update or other changes the way python
calculates the name of the file in which it stores the metadata with
which it was built, which is called _sysconfigdata_some_metadata. Don't
know which update this was or when it happened or indeed what the old
system was, but this is the name of the file that gets built for native
python in current yocto, and prevents an error about not being able to
find _sysconfigdata.py (because that error comes from a try/except meant
to catch even _older_ versions of python that didn't have anything in
the filename) and obscures the actual problem.