refactor(ot3): Set bit timings for 500kbaud
These bit timings can be achieved exactly by all microcontrollers as
well as the MCP2518, which is important because they all need to be
within 0.5%. We could also specify bit timings via (bitrate,
sample-point), but then we might generate the wrong ones so doing it
via (tq, prop-seg, phase-seg1, phase-seg2) offers more control. It also
requires a lot more settings, so reorganize the systemd service to be a
little more friendly to reading and editing.