fix(can): align mcp25xx prefetch
We've been seeing corrupted data on the canbus on the verdin every now
and again. Per
https://community.toradex.com/t/can-problem-on-verdin-i-mx8m-mini/12567/24
, that's because the driver for the canbus controller doesn't properly
ensure that prefetch amounts are word-aligned, and therefore things get
screwy. By making the prefetch word-aligned, we fix the issue. This
should be fixed with a kernel patch but for now this will work.