recipes-robot/robot-app: use system python (#57)
* recipes-robot/robot-app: do not bundle python

Since if we do, we have to play a bunch of games with cross-compiling
the python modules that then get installed. Instead, we can take
advantage of the already-installed python modules.

* recipes-robot/robot-app: use system python

The app uses python for protocol analysis. In distributed builds, it
bundles a python interpreter and the modules necessary for analysis
inside its own bundle. Here, that would be duplicative and also a pain
to build. Instead, let's tell the app to use the system interpreter, and
pass a pythonpath env var through to use the python deps in
/opt/opentrons-robot-server .